edit type tongtien

+++ database/migrations/2021_05_17_021827_create_phieu_quies_table.php
@@ -15,7 +15,7 @@ class CreatePhieuQuiesTable extends Migration
     {
         Schema::create('phieu_quies', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('nhanvien_id');
+            $table->bigInteger('nhanvien_id');
             $table->date('ngay');
             $table->text('noi_dung');
             $table->float('so_tien_thu');
